Write the function that would display numbers
Add classes of numbers to all the numbers in the
index.html and write the displayNumbers() function.
It will loop through all elements that have the class
of .number and attach a click event listener to them.
Whenever any of them are clicked, the text content of
the display (id of screen) will be the text content of the clicked number. This new text content will be concatenated
with the old text content of the display. This will
enable users to input 89 and not 8 and 9.
If they are not concatenated, when you type a number and
type another, the new number replaces the old number.

Also added the defer attribute to the script tag in the
index.html