Refactor the displayNumbers()
Rename it to handleNumbers(). Add the
click event listener to the number buttons and
if textContent exists, the textContent of the clicked
button is finally saved to secondNumber and displayed
on the screen. If an operator is clicked and certain
conditions are met, the second number is pushed into
the finalOperands array. The operator is also displayed
on the screen and pushed into the finalOperands array.

Pretty twisted things happening here.