Refactor both the functions for the operators and the handleCalculations() function
First these functions would remove the operator
sign with splice(a, b) - from index a, remove b
element. Then the functions perform the operation
on the remaining 2 elements of the array using the
array.reduce() method.

For the handleCalculations() function, if the user
clicks a number, and operator, another number and
another operator, just before adding the second
operator to the finalOperands array, it calls any
of the operating functions (depending on the operator
symbol) on the 2 numbers in the finalOperands array.
It stores the answer in the result variable and puts
it into the finalOperands array as the first element
(unshift). It then removes the previous two operands
of the array with splice(1, 2) so that the array can
be of the form
`[answerOfTheLastCalculation, operatorForNewCalculation]`
It now awaits the next operand.