Create a function to handle the finalOperands array after equals sign has been input.
The cleanUpArrayAfterEquals(operation) pushes the
secondNumber into the finalOperands, assigns the
result of an operation of the finalOperands array
to the result variable. It then makes this result
to be the first element in the finalOperands array
It then removes the previous two operands
of the array with splice(1, 2) so that the array can
be of the form
`[answerOfTheLastCalculation, operatorForNewCalculation]`
It now awaits the next operand and assigns secondNumber
to firstNumber (converted to number).

Set an event listener to the equalsButton so that if
the equalsButton.textContent exists and the length of
the finalOperands is equals to 2, like [7, "+"], it
calls the cleanUpArrayAfterEquals(operation), depending
on the operator that is the second element of the
finalOperands array (finalOperands[1]).