Fix the issue with 0 being unable to be input as firstNumber
Previously, when 0 is clicked on first, then operator
then the next operand, this last operand clicked would
be the first number as the 0 wont even be pushed into
the finalOperands array. Thus operations like
`[0, '-' 1]` could not be achieved. I fixed it by removing
the Unary plus in firstNumber and instead use the array.map()
on all the operator functions so that the textContent are pushed
into the array as strings and when the operator functions are
called, they are first converted to numbers before the operation.

https://stackoverflow.com/questions/4437916/how-to-convert-all-elements-in-an-array-to-integer-in-javascript

I read that it is because JavaScript takes `0` as a falsy value
https://stackoverflow.com/questions/60981664/javascript-how-to-convert-string-0-to-number-0