Complete work on the +/- toggle button for secondNumber and operands in general
If `result` does not exist and `firstNumber` does not
include `-`, clicking on the +/- button will add `-` to
`firstNumber`. Else (firstNumber includes `-`) clicking on
the +/- button will take the minus away from `firstNumber`.

Now if any of the operands is positive, clicking on
the +/- button will make the number negative. And if
negative, will make the number positive. The perfect
toggle button.