Fix the bug with the +/- toggle button and results
Previously, after a result is given, the +/- button
would not work on the next operand (result of the
previous calculation with a new operand). Had the
calculator check if the length of the finalOperands
array was greater than one and there was an operator.

Could not fix this issue for the decimal button.
Now if the result is given and an operator is clicked
and you try to use the decimal button to make the next
operand a decimal, it would not work.