Add the feature that throws an error that the user cannot divide by 0 in a string of calculations
In the division() function, if array[1] is 0 (secondNumber),
it will return an error.

To make the calculator not add (array.unshift) the
error message which is the result and the next operator
to the finalOperands array, these are performed if
the result is not the error. Else it empties the
finalOperands array.

These happens for a string of continuous calculations

Yet to implement it for when the equals button is used.