Add the feature that throws an error that the user cannot divide by 0 in an `=` situation.
When a user tries to divide by 0 and uses `=`,
To make the calculator not add (array.unshift) the
error message which is the result and the next operator
to the finalOperands array, these are performed if
the result is not the error. Else it empties the
finalOperands array. Implemented this is the
cleanUpArrayAfterEquals(operation) function.

After this, to prevent the 0
(the secondNumber from [7, "/", 0]) from appending
to the next secondNumber, I tried to empty first and
second numbers in lines 239 and 240 but emptying
them would prevent users from making a continuous
string of calculations after the `=` is used