Start adding the keyboard support functionality
Create a `keydown` event listener on the document
object so that when a user presses a key on their
keyboard, it will capture the `event.key`.

Had to create functions to handle the input of numbers,
decimals and the +/- button. It is now convenient to
call these functions when the user clicks on the buttons
concerned and more recently, when the user presses
a number or `.` key on their keyboard.

While creating a function for the +/- button, I noticed
that when trying to use the button to toggle the sign
of the result (after using equals), it would throw a
`Reference Error: negativeButtonText is not defined`
but still works as expected. If i change it to
`theInput` (as it should rightly be) or "-", the feature
will break and display the secondNumber instead of the result
as the user tries to toggle the sign. It will display
the secondNumber but still work with the result

Will fix this.