Continue styling the row, zero and buttons
Change the last-row class in the index.html to
row and give the row a justify-content of
space-around. Give the buttons width, height
and border-radius. Give the #zero button a
different width and a different border radius.

The % value border radius worked for other buttons
to give them the circle shape but I had to use
px border radius value for the 0 button to give it
that shape

C'est parti!