Style the screen so that the display is aligned to the right
Give screen a `justify-content` property of `right`
to take the display to the right and an `align-item`
property of `flex-end` to align it to bottom.

Give the calc-container and screen a padding of 10px
each to add more space around the children and the
screen respectively. This makes the screen display
to be aligned to the last buttons on the left.The
justify-content and align-items properties that were
given to the screen display caused it to to aligned
without spaces to the calc-container