Add a pressed down and hover effect on the buttons
In javascript.js add a click event listener to the buttons. On click the function will toggle
the .active class on the button that was clicked.
I also gave the class properties un the styles.css. This gives the pressed effect. A setTimeout
is also set on the button that was clicked to remove the .active class after 1000 milliseconds (1 second). This makes the button give the effect of coming back up.

More on this at `https://dev.to/nicm42/how-to-make-a-button-looked-like-it-s-staying-pressed-down-58k` and `https://codepen.io/nicm42/pen/OJjdoqe`

Also give the buttons a pseudo class of :hover.
Upon hover, the background color changes according
to the transition-duration property in the button
style.